requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
python-dateutil==2.9.0.post0
pandas==2.2.2
gspread
google-auth
